Update DST Rework
Existing logic was wrong. It did fix the specific issue given we are IN DST right now, but didn't fix the whole problem (year round) because $dst = $tz.IsDaylightSavingTime(0) never returns the correct boolean.
IsDaylightSavingTime does accurately report a boolean from the Get-Date command. So this new logic is if supportDst=True AND IsDaylightSavingTime=true then drop an hour.